tamanho máximo do grafo limitado a 46000 enquanto não percebo o que se passa com o cálculo da memória a alocar. Adicionados comentários e printfs para debug

+++ src/graph/graph.h
@@ -36,7 +36,7 @@ void create_graph(int graph_size, int edge_percentage);
 void create_locked_graph(int graph_size, int edge_percentage);
 void add_edge(float *graph, int u, int v, float weight);
 void add_null_edge(float *graph, int u, int v);
-unsigned int get_matrix_size(int graph_size);
+unsigned long int get_matrix_size(int graph_size);
 
 float get_edge(float *graph, int u, int v);
 int get_edge_count();
